Pybytes 1.3.0 (#372)
Pybytes scripts REL 1.3.0
modpycom: New sigfox_info() API

+++ esp32/frozen/Pybytes/_OTA.py
@@ -5,6 +5,20 @@ later version, with permitted additional terms. For more information
 see the Pycom Licence v1.0 document supplied with this file, or
 available at https://www.pycom.io/opensource/licensing
 '''
+try:
+    from pybytes_debug import print_debug
+except:
+    from _pybytes_debug import print_debug
+
+try:
+    from pybytes_constants import constants
+except:
+    from _pybytes_constants import constants
+
+try:
+    import urequest
+except:
+    import _urequest as urequest
 
 import network
 import socket
@@ -17,11 +31,6 @@ import gc
 import pycom
 import os
 from binascii import hexlify
-
-try:
-    from pybytes_debug import print_debug
-except:
-    from _pybytes_debug import print_debug
 
 # Try to get version number
 # try:
@@ -40,6 +49,9 @@ class OTA():
     def get_data(self, req, dest_path=None, hash=False):
         raise NotImplementedError()
 
+    def update_device_network_config(self, fcota, config):
+        raise NotImplementedError()
+
     # OTA methods
 
     def get_current_version(self):
@@ -56,9 +68,9 @@ class OTA():
         gc.collect()
         return manifest
 
-    def update(self):
+    def update(self, customManifest=None):
         try:
-            manifest = self.get_update_manifest()
+            manifest = self.get_update_manifest() if not customManifest else customManifest
         except Exception as e:
             print('Error reading the manifest, aborting: {}'.format(e))
             return 0
@@ -68,36 +80,38 @@ class OTA():
             return 1
 
         # Download new files and verify hashes
-        for f in manifest['new'] + manifest['update']:
-            # Upto 5 retries
-            for _ in range(5):
-                try:
-                    self.get_file(f)
-                    break
-                except Exception as e:
-                    print(e)
-                    msg = "Error downloading `{}` retrying..."
-                    print(msg.format(f['URL']))
-                    return 0
-            else:
-                raise Exception("Failed to download `{}`".format(f['URL']))
+        if "new" in manifest and "update" in manifest:
+            for f in manifest['new'] + manifest['update']:
+                # Upto 5 retries
+                for _ in range(5):
+                    try:
+                        self.get_file(f)
+                        break
+                    except Exception as e:
+                        print(e)
+                        msg = "Error downloading `{}` retrying..."
+                        print(msg.format(f['URL']))
+                        return 0
+                else:
+                    raise Exception("Failed to download `{}`".format(f['URL']))
 
-        # Backup old files
-        # only once all files have been successfully downloaded
-        for f in manifest['update']:
-            self.backup_file(f)
+            # Backup old files
+            # only once all files have been successfully downloaded
+            for f in manifest['update']:
+                self.backup_file(f)
 
-        # Rename new files to proper name
-        for f in manifest['new'] + manifest['update']:
-            new_path = "{}.new".format(f['dst_path'])
-            dest_path = "{}".format(f['dst_path'])
+            # Rename new files to proper name
+            for f in manifest['new'] + manifest['update']:
+                new_path = "{}.new".format(f['dst_path'])
+                dest_path = "{}".format(f['dst_path'])
 
-            os.rename(new_path, dest_path)
+                os.rename(new_path, dest_path)
 
-        # `Delete` files no longer required
-        # This actually makes a backup of the files incase we need to roll back
-        for f in manifest['delete']:
-            self.delete_file(f)
+        if "delete" in manifest:
+            # `Delete` files no longer required
+            # This actually makes a backup of the files incase we need to roll back
+            for f in manifest['delete']:
+                self.delete_file(f)
 
         # Flash firmware
         if "firmware" in manifest:
@@ -162,9 +176,12 @@ class OTA():
         os.rename(dest_path, bak_path)
 
     def write_firmware(self, f):
-        hash = self.get_data(f['URL'].split("/", 3)[-1],
-                             hash=True,
-                             firmware=True)
+        # hash =
+        self.get_data(
+            f['URL'].split("/", 3)[-1],
+            hash=True,
+            firmware=True
+        )
         # TODO: Add verification when released in future firmware
 
 
@@ -258,6 +275,7 @@ class WiFiOTA(OTA):
             if fp is not None:
                 fp.close()
             if firmware:
+                print_debug(6, 'ota_finish')
                 pycom.ota_finish()
 
         except Exception as e:
@@ -277,3 +295,47 @@ class WiFiOTA(OTA):
                 return bytes(content)
         elif hash:
             return hash_val
+
+    def update_device_network_config(self, fcota, config):
+        targetURL = '{}://{}/device/networks/{}'.format(
+            constants.__DEFAULT_PYCONFIG_PROTOCOL, constants.__DEFAULT_PYCONFIG_DOMAIN, config['device_id']
+        )
+        print_debug(6, "request device update URL: {}".format(targetURL))
+        try:
+            pybytes_activation = urequest.get(targetURL, headers={'content-type': 'application/json'})
+            responseDetails = pybytes_activation.json()
+            pybytes_activation.close()
+            print_debug(6, "Response Details: {}".format(responseDetails))
+            self.update_network_config(responseDetails, fcota, config)
+            machine.reset()
+        except Exception as ex:
+            print_debug(1, "error while calling {}!: {}".format(targetURL, ex))
+
+    def update_network_config(self, letResp, fcota, config):
+        try:
+            if 'networkConfig' in letResp:
+                netConf = letResp['networkConfig']
+                config['network_preferences'] = netConf['networkPreferences']
+                if 'wifi' in netConf:
+                    config['wifi'] = netConf['wifi']
+                elif 'wifi' in config:
+                    del config['wifi']
+
+                if 'lte' in netConf:
+                    config['lte'] = netConf['lte']
+                elif 'lte' in config:
+                    del config['lte']
+
+                if 'lora' in netConf:
+                    config['lora'] = {
+                        'otaa': netConf['lora']['otaa'],
+                        'abp': netConf['lora']['abp']
+                    }
+                elif 'lora' in config:
+                    del config['lora']
+
+                json_string = ujson.dumps(config)
+                print_debug(1, "update_network_config : {}".format(json_string))
+                fcota.update_file_content('/flash/pybytes_config.json', json_string)
+        except Exception as e:
+            print_debug(1, "error while updating network config pybytes_config.json! {}".format(e))
